refactor(component): move getJsonDoc to service dir
